+++ mtpc-basics/src/test/java/dev/kennethgjohnson/mtpc/basics/concurrency/reentrantlocks/ReentrantProduceConsumerWithWaitSignalConditionsAppTest.java
@@ -1,0 +1,14 @@
+package dev.kennethgjohnson.mtpc.basics.concurrency.reentrantlocks;
+
+import static org.junit.jupiter.api.Assertions.assertTrue;
+import org.junit.jupiter.api.Test;
+
+public class ReentrantProduceConsumerWithWaitSignalConditionsAppTest {
+
+  @Test
+  public void shouldBeAbleToExecuteMainWithoutException() {
+    String[] arguments = {};
+    ReentrantProduceConsumerWithWaitSignalConditionsApp.main(arguments);
+    assertTrue(true);
+  }
+}
